refactor!: relocate support classes
delete unreferenced classes
