+++ commons/src/main/java/org/smooks/xml/XsdDOMValidator.java
@@ -44,7 +44,7 @@ package org.smooks.xml;
 
 import org.smooks.api.SmooksException;
 import org.smooks.assertion.AssertArgument;
-import org.smooks.io.StreamUtils;
+import org.smooks.support.StreamUtils;
 import org.smooks.support.ClassUtil;
 import org.w3c.dom.*;
 import org.w3c.dom.ls.LSInput;
