+++ commons/src/test/java/org/smooks/archive/ArchiveTestCase.java
@@ -54,7 +54,7 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
 
-import org.smooks.io.StreamUtils;
+import org.smooks.support.StreamUtils;
 
 /**
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
