+++ commons/src/test/java/org/smooks/support/XmlUtilTestCase.java
@@ -54,7 +54,6 @@ import java.io.Reader;
 import java.io.StringReader;
 import java.io.StringWriter;
 
-import org.smooks.io.StreamUtils;
 import org.smooks.xml.LocalDTDEntityResolver;
 import org.w3c.dom.Document;
 import org.xml.sax.SAXException;
