+++ core/src/test/java/org/smooks/engine/delivery/event/ExecutionReportGeneratorTestCase.java
@@ -49,7 +49,7 @@ import org.smooks.Smooks;
 import org.smooks.api.ExecutionContext;
 import org.smooks.engine.report.FlatReportGenerator;
 import org.smooks.engine.report.ReportConfiguration;
-import org.smooks.io.StreamUtils;
+import org.smooks.support.StreamUtils;
 import org.xml.sax.SAXException;
 
 import javax.xml.transform.stream.StreamResult;
