+++ core/src/test/java/org/smooks/engine/delivery/sax/SmooksSAXFilterTestCase.java
@@ -56,7 +56,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.smooks.Smooks;
 import org.smooks.api.ExecutionContext;
-import org.smooks.io.StreamUtils;
+import org.smooks.support.StreamUtils;
 import org.xml.sax.SAXException;
 
 /**
